Rename jc_graph.py to f4a_graph.py and update all import references
