+++ tests/test_job_loading.py
@@ -6,7 +6,7 @@ import pytest
 import yaml
 
 from flow4ai import f4a_logging as logging
-from flow4ai.jc_graph import validate_graph
+from flow4ai.f4a_graph import validate_graph
 from flow4ai.job import JobABC, Task, job_graph_context_manager
 from flow4ai.job_chain import JobChain  # Import JobChain
 from flow4ai.job_loader import ConfigLoader, ConfigurationError, JobFactory
